% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_balance_test.R
\name{compute_balance_test}
\alias{compute_balance_test}
\title{Wald-test if Xhat is statistically different from zero}
\usage{
compute_balance_test(Ybar_g_list, A_0_list, S_g_list, N_g_list)
}
\arguments{
\item{Ybar_g_list}{Ybar_g_list}

\item{A_0_list}{A_0_list}

\item{S_g_list}{S_g_list}

\item{N_g_list}{N_g_list}
}
\value{
Wald_test_Xhat Wald-test statistic for the balance test.
}
\description{
\code{balance_check} computes the Wald test-statistic (F-test) for the null that E[X]=0
}
