% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_efficient_estimator_and_se.R
\name{compute_Betastar}
\alias{compute_Betastar}
\title{Plug-in efficient Beta hat}
\usage{
compute_Betastar(
  Ybar_g_list,
  A_theta_list,
  A_0_list,
  S_g_list,
  N_g_list,
  Xvar_list = NULL
)
}
\description{
\code{compute_Betastar} computes the plug-in efficient betahat
}
