% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_efficient_estimator_and_se.R
\name{compute_Xhat}
\alias{compute_Xhat}
\title{Compute Xhat of pre-treatment differences}
\usage{
compute_Xhat(Ybar_g_list, A_0_list)
}
\arguments{
\item{Ybar_g_list}{Ybar_g_list}

\item{A_0_list}{A_0_list}
}
\value{
Xhat the vector Xhat of pre-treatment differences to be used as regressors
}
\description{
\code{compute_Xhat} computes the vector Xhat of pre-treatment differences given the list of cohort means
Ybar_g_list and the list of matrices A_0_list
}
