% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_efficient_estimator_and_se.R
\name{create_A0_list}
\alias{create_A0_list}
\title{create_A0_list}
\usage{
create_A0_list(g_list, t_list)
}
\arguments{
\item{g_list}{g_list}

\item{t_list}{t_list}
}
\value{
A0_list list of A_0 matrices for Xhat corresponding with all possible comparisons of cohorts before they are treated
}
\description{
\code{create_A0_list} creates the list of A_0 matrices for Xhat corresponding with all possible
comparisons of cohorts before they are treated
}
